Add new auth components for SignIn, SignUp, ResetPassword. Redirect signed in users to home page on those component based pages.
